+++ src/components/OrdersStatus/OrderColumn/OrdersColumn.module.css
@@ -1,0 +1,4 @@
+.orders_list {
+  margin: 0;
+  padding: 0;
+}
