+++ src/components/OrdersStatus/OrdersStatus.module.css
@@ -10,16 +10,11 @@
     4px 0px 8px rgba(51, 51, 255, 0.25), 0px 4px 32px rgba(51, 51, 255, 0.5);
 }
 
-.orders_list {
-  margin: 0;
-  padding: 0;
-}
-
 .numbers_container {
   display: flex;
   flex-direction: row;
   justify-content: flex-start;
-  height: 300px;
+  height: 200px;
   margin: 0;
   padding: 0;
   margin-top: 12px;
