+++ src/data/ru.json
@@ -38,7 +38,11 @@
   "OrderStatus": {
     "done": "Выполнен",
     "created": "Создан",
-    "pending": "Готовится"
+    "pending": "Готовится",
+    "all": "Выполнено за все время:",
+    "today": "Выполнено за сегодня:",
+    "ready": "Готовы:",
+    "process": "В работе:"
   },
   "IngredientDetails": {
     "title": "Детали ингредиента",
